Audio/ValKit: tdAudioTest.py: Log stuff in killHstProcessByName() to find out what's going on. bugref:10008
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@90754 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/ValidationKit/tests/audio/tdAudioTest.py
@@ -279,16 +279,18 @@ class tdAudioTest(vbox.TestDriver):
             # Using the BSD syntax here; MacOS also should understand this.
             procPs = subprocess.Popen(['ps', 'ax'], stdout=subprocess.PIPE);
             out, err = procPs.communicate();
-            reporter.log2('PS stderr:');
-            for sLine in err.decode("utf-8").splitlines():
-                reporter.log2(sLine);
-            reporter.log2('PS stdout:');
-            for sLine in out.decode("utf-8").splitlines():
-                reporter.log2(sLine);
-                if sProcName in sLine:
-                    pid = int(sLine.split(None, 1)[0]);
-                    reporter.log('Killing PID %d' % (pid,));
-                    os.kill(pid, signal.SIGKILL); # pylint: disable=no-member
+            if err:
+                reporter.log2('PS stderr:');
+                for sLine in err.decode("utf-8").splitlines():
+                    reporter.log2(sLine);
+            if out:
+                reporter.log2('PS stdout:');
+                for sLine in out.decode("utf-8").splitlines():
+                    reporter.log2(sLine);
+                    if sProcName in sLine:
+                        pid = int(sLine.split(None, 1)[0]);
+                        reporter.log('Killing PID %d' % (pid,));
+                        os.kill(pid, signal.SIGKILL); # pylint: disable=no-member
 
     def killHstVkat(self):
         """
